Moved service methods to WeatherServiceImpl instead of WeatherRepositoryImpl
